move core beacon functionality to a core package
